Update readme and other files
