+++ app.json
@@ -1,7 +1,7 @@
 {
-  "name": "Asylum Connect Catalog",
+  "name": "AsylumConnect Catalog",
   "description": "The updated Asylum Connect Catalog",
-  "repository": "https://github.com/jeremyhatter/asylum-connect-1deg-catalog",
+  "repository": "https://github.com/asylum-connect/asylumconnect-catalog",
   "logo": "https://cdn.rawgit.com/heroku/node-js-getting-started/master/public/node.svg",
   "keywords": ["node", "react", "heroku"],
   "image": "heroku/nodejs"
