+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "asylum-connect-catalog",
+  "name": "asylumconnect-catalog",
   "version": "0.1.0",
   "description": "Asylum Connect Catalog App",
   "engines": {
