--- src/locale/en_US.js
@@ -1,1 +1,0 @@
-export default {};
