--- src/theme/breakpoints.js
@@ -1,9 +1,0 @@
-const breakpoints = {
-  xs: 360,
-  sm: 600,
-  md: 960,
-  lg: 1280,
-  xl: 1500
-};
-
-export default breakpoints;
